The Lookup class wasn't pulling its weight.
The Lookup class could be moved into the Dispatch class directly.
This gets rid of another indirection. The Dispatch objects now
directly address key_lookup.
