+++ reg/predicate.py
@@ -446,145 +446,6 @@ class CachingKeyLookup(object):
         return result
 
 
-class Lookup(object):
-    """The lookup is used for generic dispatch.
-
-    The arguments to the lookup functions are those of the dispatch
-    function being called. The lookup extracts the predicate_key from
-    these arguments and then looks up the actual function to call. This
-    function is then called with the original arguments.
-
-    :param key_lookup: the key lookup, either a :class:`PredicateRegistry` or
-      :class:`CachingKeyLookup`.
-    """
-    def __init__(self, callable, key_lookup):
-        self.key_lookup = key_lookup
-        self.callable = callable
-        self.arg_extractor = ArgExtractor(callable, self.key_lookup.argnames())
-
-    def predicate_key(self, *args, **kw):
-        """Construct predicate_key for function arguments.
-
-        For function arguments, construct the appropriate
-        ``predicate_key``. This is used by the dispatch mechanism to
-        dispatch to the right function.
-
-        If the ``predicate_key`` cannot be constructed from ``args``
-        and ``kw``, this raises a :exc:`KeyExtractorError`.
-
-        :param args: the varargs given to the callable.
-        :param kw: the keyword arguments given to the callable.
-        :returns: an immutable ``predicate_key`` based on the predicates
-          the callable was configured with.
-        """
-        return self.key_lookup.key(self.arg_extractor(*args, **kw))
-
-    def call(self, *args, **kw):
-        """Call with args and kw.
-
-        If nothing more specific is registered, call the dispatch
-        function as a fallback.
-
-        :args: varargs for the call. Is also used to extract
-           dispatch information to construct predicate_key.
-        :kw: keyword arguments for the call. Is also used to extract
-           dispatch information to construct predicate_key.
-        :returns: the result of the call.
-        """
-        try:
-            key = self.predicate_key(*args, **kw)
-            component = self.key_lookup.component(key)
-        except KeyExtractorError:
-            # if we cannot extract the key this could be because the
-            # dispatch was never initialized, as register_function
-            # was not called. In this case we want the fallback.
-            # Alternatively we cannot construct the key because we
-            # were passed the wrong arguments.
-            # In both cases, call the fallback. In case the wrong arguments
-            # were passed, we get the appropriate TypeError then
-            component = self.callable
-
-        if component is None:
-            # try to use the fallback
-            component = self.key_lookup.fallback(key)
-            if component is None:
-                # if fallback is None use the original callable as fallback
-                component = self.callable
-        return component(*args, **kw)
-
-    def component(self, *args, **kw):
-        """Lookup function dispatched to with args and kw.
-
-        Looks up the function to dispatch to using args and
-        kw. Returns the fallback value (default: ``None``) if nothing
-        could be found.
-
-        :args: varargs. Used to extract dispatch information to
-           construct ``predicate_key``.
-        :kw: keyword arguments. Used to extract
-           dispatch information to construct ``predicate_key``.
-        :returns: the function being dispatched to, or None.
-        """
-        key = self.predicate_key(*args, **kw)
-        return self.key_lookup.component(key)
-
-    def fallback(self, *args, **kw):
-        """Lookup fallback for args and kw.
-
-        :args: varargs. Used to extract dispatch information to
-           construct ``predicate_key``.
-        :kw: keyword arguments. Used to extract
-           dispatch information to construct ``predicate_key``.
-        :returns: the function being dispatched to, or fallback.
-        """
-        key = self.predicate_key(*args, **kw)
-        return self.key_lookup.fallback(key)
-
-    def component_by_keys(self, **kw):
-        """Look up function based on key_dict.
-
-        Looks up the function to dispatch to using a key_dict,
-        mapping predicate name to predicate value. Returns the fallback
-        value (default: ``None``) if nothing could be found.
-
-        :kw: key is predicate name, value is
-          predicate value under which it was registered.
-          If omitted, predicate default is used.
-        :returns: the function being dispatched to, or fallback.
-        """
-        key = self.key_lookup.key_dict_to_predicate_key(kw)
-        return self.key_lookup.component(key)
-
-    def all(self, *args, **kw):
-        """Lookup all functions dispatched to with args and kw.
-
-        Looks up functions for all permutations based on predicate_key,
-        where predicate_key is constructed from args and kw.
-
-        :args: varargs. Used to extract dispatch information to
-           construct predicate_key.
-        :kw: keyword arguments. Used to extract
-           dispatch information to construct predicate_key.
-        :returns: an iterable of functions.
-        """
-        key = self.predicate_key(*args, **kw)
-        return self.key_lookup.all(key)
-
-    def all_key_dict(self, **key_dict):
-        """Look up all functions dispatched to using on key_dict.
-
-        Looks up the function to dispatch to using a ``key_dict``,
-        mapping predicate name to predicate value. Returns the fallback
-        value (default: ``None``) if nothing could be found.
-
-        :key_dict: a dictionary. key is predicate name, value is
-          predicate value. If omitted, predicate default is used.
-        :returns: iterable of functions being dispatched to.
-        """
-        key = self.key_lookup.key_dict_to_predicate_key(key_dict)
-        return self.key_lookup.all(key)
-
-
 # XXX transform to non-recursive version
 # use # http://blog.moertel.com/posts/2013-05-14-recursive-to-iterative-2.html
 def multipredicate_permutations(indexes, keys):
